[IMP] Moved endpoints in namespace app.blueprints.rest.v2
