+++ source/app/views.py
@@ -99,6 +99,7 @@ from app.blueprints.rest.search_routes import search_rest_blueprint
 from app.blueprints.graphql.graphql_route import graphql_blueprint
 from app.blueprints.rest.v2.case.api_v2_assets_routes import api_v2_assets_blueprint
 from app.blueprints.rest.v2.case.api_v2_ioc_routes import api_v2_ioc_blueprint
+from app.blueprints.rest.v2.case.api_v2_case_tasks_routes import api_v2_tasks_blueprint
 from app.models.authorization import User
 from app.post_init import run_post_init
 
@@ -186,6 +187,7 @@ app.register_blueprint(demo_blueprint)
 
 app.register_blueprint(api_v2_ioc_blueprint)
 app.register_blueprint(api_v2_assets_blueprint)
+app.register_blueprint(api_v2_tasks_blueprint)
 
 try:
